Handle LOAD_PATH similarly to how MRI does it
This patch changes the way Rubinius currently handles the $LOAD_PATH and
some related issues to make it behave more like MRI on that front, and
thus making it easier to port existing Ruby infrastructure to use with
Rubinius.

Look how the $LOAD_PATH works in MRI, commented:

$ ruby1.8 -e 'puts $:'
/usr/local/lib/site_ruby/1.8                  # sitelibdir
/usr/local/lib/site_ruby/1.8/x86_64-linux     # sitearchdir
/usr/local/lib/site_ruby                      # sitedir
/usr/lib/ruby/vendor_ruby/1.8                 # vendorlibdir
/usr/lib/ruby/vendor_ruby/1.8/x86_64-linux    # vendorarchdir
/usr/lib/ruby/vendor_ruby                     # vendordir
/usr/lib/ruby/1.8                             # rubylibdir
/usr/lib/ruby/1.8/x86_64-linux                # archdir
.                                             # current dir :-p
$ ruby1.9.1 -e 'puts $:'
/usr/local/lib/site_ruby/1.9.1                # sitelibdir
/usr/local/lib/site_ruby/1.9.1/x86_64-linux   # sitearchdir
/usr/local/lib/site_ruby                      # sitedir
/usr/lib/ruby/vendor_ruby/1.9.1               # vendorlibdir
/usr/lib/ruby/vendor_ruby/1.9.1/x86_64-linux  # vendorarchdir
/usr/lib/ruby/vendor_ruby                     # vendordir
/usr/lib/ruby/1.9.1                           # rubylibdir
/usr/lib/ruby/1.9.1/x86_64-linux              # archdir

Apart from 1.9 dropping the current directory (".") from the $LOAD_PATH, they
are pretty much the same, varying only by the version name. A nice property of
this scheme is that ${sitedir} and ${vendordir} are in the $LOAD_PATH for both
1.8 and 1.9.

In Debian, we are migrating all our Ruby packages to use this fact and
install all pure-Ruby software (*.rb) that is known to work in both 1.8
and 1.9 to ${vendordir} so that they are available to all supported Ruby
versions.  Packages with compiled extensions install the pure Ruby code
to ${vendordir} and the compiled extensions for each MRI version to the
corresponding ${vendorarchdir}, so that with a single package we support
both MRI versions. Packages that are by design only for a specific
version (e.g. rcov) can be installed to the corresponding
${vendorlibdir} and thus will only be available for that version.

My plan is to add Rubinius to this scheme, so that once it gets in, all
pure-Ruby software will work out of the box with Rubinius. The packages
with C extensions will only need a recompilation once after our
packaging helper tool gets updated to include Rubinius in the build
process.

So this patch is my proposal to make Rubinius handle the $LOAD_PATH in a way
that is more similar to how MRI does it.

The entries ${vendordor,sitedir}/${arch} are being dropped. This will
break users who have installed packages by hand (i.e.  setup.rb or
extconf.rb, not using Rubygems). This doesn't seem to be a common usage
pattern in the Ruby community, though; but you probably know more about
that and can comment.